Moved configuration handling to its own service
